chore: handle PayForBlobNamespaceID in `merge` (#1298)
Closes https://github.com/celestiaorg/celestia-app/issues/1285

## Description
1. Handle PayForBlobNamespaceID in `merge`
1. Extract `merge` to a new file
1. Extract `merge` to a new file
1. Add a unit test that uses sample block data
