+++ pkg/shares/shares_test.go
@@ -2,67 +2,14 @@ package shares
 
 import (
 	"bytes"
-	"context"
 	"testing"
-	"time"
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
-	"github.com/celestiaorg/celestia-app/testutil/testfactory"
-	"github.com/celestiaorg/rsmt2d"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 	coretypes "github.com/tendermint/tendermint/types"
 )
-
-func TestMerge(t *testing.T) {
-	type test struct {
-		name      string
-		txCount   int
-		blobCount int
-		maxSize   int // max size of each tx or blob
-	}
-
-	tests := []test{
-		{"one of each random small size", 1, 1, 40},
-		{"one of each random large size", 1, 1, 400},
-		{"many of each random large size", 10, 10, 40},
-		{"many of each random large size", 10, 10, 400},
-		{"only transactions", 10, 0, 400},
-		{"only blobs", 0, 10, 400},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			// generate random data
-			data := generateRandomBlockData(
-				tc.txCount,
-				tc.blobCount,
-				tc.maxSize,
-			)
-
-			shares, err := Split(data, false)
-			require.NoError(t, err)
-			rawShares := ToBytes(shares)
-
-			eds, err := rsmt2d.ComputeExtendedDataSquare(rawShares, appconsts.DefaultCodec(), rsmt2d.NewDefaultTree)
-			if err != nil {
-				t.Error(err)
-			}
-
-			res, err := merge(eds)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			res.SquareSize = data.SquareSize
-
-			assert.Equal(t, data, res)
-		})
-	}
-}
 
 // TestPadFirstIndexedBlob ensures that we are adding padding to the first share
 // instead of calculating the value.
@@ -88,43 +35,6 @@ func TestPadFirstIndexedBlob(t *testing.T) {
 	require.NoError(t, err)
 
 	require.True(t, bytes.Contains(resShare, blob))
-}
-
-func TestFuzz_Merge(t *testing.T) {
-	t.Skip()
-	// run random shares through processCompactShares for a minute
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			TestMerge(t)
-		}
-	}
-}
-
-// generateRandomBlockData returns randomly generated block data for testing purposes
-func generateRandomBlockData(txCount, blobCount, maxSize int) (data coretypes.Data) {
-	data.Txs = testfactory.GenerateRandomlySizedTxs(txCount, maxSize)
-	data.Blobs = testfactory.GenerateRandomlySizedBlobs(blobCount, maxSize)
-	data.SquareSize = appconsts.DefaultMaxSquareSize
-	return data
-}
-
-// generateRandomBlobOfShareCount returns a blob that spans the given
-// number of shares
-func generateRandomBlobOfShareCount(count int) coretypes.Blob {
-	size := rawBlobSize(appconsts.FirstSparseShareContentSize * count)
-	return testfactory.GenerateRandomBlob(size)
-}
-
-// rawBlobSize returns the raw blob size that can be used to construct a
-// blob of totalSize bytes. This function is useful in tests to account for
-// the delimiter length that is prefixed to a blob's data.
-func rawBlobSize(totalSize int) int {
-	return totalSize - DelimLen(uint64(totalSize))
 }
 
 func TestSequenceLen(t *testing.T) {
